+++ packages/plugins/openid4vc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @learncard/openid4vc-plugin
 
+## 0.3.9
+
+### Patch Changes
+
+- Updated dependencies [[`0c1bf9a8a33e6392d5fd279479d9ab4fb0449b5e`](https://github.com/learningeconomy/LearnCard/commit/0c1bf9a8a33e6392d5fd279479d9ab4fb0449b5e)]:
+    - @learncard/vc-plugin@1.6.2
+    - @learncard/core@9.4.36
+    - @learncard/didkit-plugin@1.10.1
+
 ## 0.3.8
 
 ### Patch Changes